Rename abstract runner to runner
