+++ gateway/api/use_cases/jobs/provider_logs.py
@@ -14,7 +14,7 @@ from api.domain.exceptions.invalid_access_exception import InvalidAccessExceptio
 from core.domain.filter_logs import filter_logs_with_non_public_tags
 from core.models import Job
 from core.utils import check_logs
-from core.services.runners import get_runner, RunnerError
+from core.services.runners.runner import Runner, RunnerError
 from core.services.storage.logs_storage import LogsStorage
 
 logger = logging.getLogger("api.GetProviderJobLogsUseCase")
@@ -50,7 +50,7 @@ class GetProviderJobLogsUseCase:
         if logs:
             return logs
 
-        runner = get_runner(job)
+        runner = Runner.get(job)
         if runner.is_active():
             try:
                 logs = runner.provider_logs()
