+++ gateway/core/services/runners/ray_runner.py
@@ -25,14 +25,14 @@ from opentelemetry import trace
 from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
 
 from core.models import ComputeResource, Job, JobConfig, DEFAULT_PROGRAM_ENTRYPOINT
-from core.services.runners.abstract_runner import AbstractRunner, RunnerError
+from core.services.runners.runner import Runner, RunnerError
 from core.services.storage.file_storage import FileStorage, WorkingDir
 from core.utils import retry_function, decrypt_env_vars, sanitize_file_path
 
 logger = logging.getLogger("RayRunner")
 
 
-class RayRunner(AbstractRunner):
+class RayRunner(Runner):
     """Client for executing jobs on Ray/KubeRay clusters."""
 
     def __init__(self, job: Job):
