+++ gateway/scheduler/schedule.py
@@ -14,7 +14,7 @@ from django.db.models.aggregates import Count, Min
 from opentelemetry import trace
 
 from core.models import Job, JobEvent, Program
-from core.services.runners import get_runner, RunnerError
+from core.services.runners.runner import Runner, RunnerError
 
 User: Model = get_user_model()
 logger = logging.getLogger("scheduler.schedule")
@@ -34,7 +34,7 @@ def execute_job(job: Job) -> Job:
     """
     tracer = trace.get_tracer("scheduler.tracer")
     with tracer.start_as_current_span("execute.job") as span:
-        runner = get_runner(job)
+        runner = Runner.get(job)
 
         try:
             runner.submit()
@@ -131,7 +131,7 @@ def fail_job_insufficient_resources(job: Job):
             job.compute_resource.title,
         )
     else:
-        runner = get_runner(job)
+        runner = Runner.get(job)
         try:
             runner.free_resources()
         except RunnerError:
