+++ gateway/tests/scheduler/management/test_commands.py
@@ -12,8 +12,8 @@ from unittest.mock import patch, MagicMock
 from prometheus_client import CollectorRegistry
 
 from core.model_managers.job_events import JobEventContext, JobEventOrigin, JobEventType
-from core.models import ComputeResource, Job, JobEvent, Config
-from core.services.runners import RunnerError
+from core.models import ComputeResource, Job, JobEvent, Program, Provider, Config
+from core.services.runners.runner import RunnerError
 from core.utils import check_logs
 from scheduler.kill_signal import KillSignal
 from scheduler.metrics.scheduler_metrics_collector import SchedulerMetrics
@@ -44,7 +44,7 @@ class TestCommands:
         num_resources = ComputeResource.objects.count()
         assert num_resources == 1
 
-    @patch("scheduler.tasks.update_jobs_statuses.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_update_jobs_statuses(self, get_runner):
         """Tests update of job statuses."""
         # Test status change from PENDING to RUNNING
@@ -187,7 +187,7 @@ class TestCommands:
         )
         assert "AAAAAAAAAAB" in logs
 
-    @patch("scheduler.tasks.update_jobs_statuses.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_update_jobs_statuses_filters_logs_user_function(self, get_runner, settings):
         """Tests that logs are filtered when saving for function without provider."""
         compute_resource = TestUtils.get_or_create_compute_resource(title="test-cluster-user-logs", active=True)
@@ -255,7 +255,7 @@ INFO: Final public log
         job.refresh_from_db()
         assert job.logs == ""
 
-    @patch("scheduler.tasks.update_jobs_statuses.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_update_jobs_statuses_filters_logs_provider_function(self, get_runner, settings):
         """Tests that logs are filtered when saving for function with provider."""
         compute_resource = TestUtils.get_or_create_compute_resource(title="test-cluster-provider-logs", active=True)
@@ -323,7 +323,7 @@ WARNING: Private warning
             saved_provider_logs = log_file.read()
         assert saved_provider_logs == expected_provider_logs
 
-    @patch("scheduler.tasks.update_jobs_statuses.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_update_jobs_statuses_job_handler_status_error_status_event(self, get_runner, settings):
         """Tests that the job_event is stored when runner.status() raises exception."""
         compute_resource = TestUtils.get_or_create_compute_resource(title="test-cluster-provider-logs", active=True)
