+++ gateway/tests/scheduler/test_schedule.py
@@ -10,7 +10,7 @@ from ray.dashboard.modules.job.common import JobStatus
 from rest_framework.test import APITestCase
 
 from core.models import Job, ComputeResource, JobEvent
-from core.services.runners import RunnerError
+from core.services.runners.runner import RunnerError
 from core.services.storage.logs_storage import LogsStorage
 
 from scheduler.kill_signal import KillSignal
@@ -71,7 +71,7 @@ class TestScheduleApi(APITestCase):
         assert str(job1.id) in job_ids  # `test4_user` job
         assert str(job6.id) in job_ids  # `test_user` job
 
-    @patch("scheduler.schedule.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_execute_job_success(self, mock_get_runner_client):
         """Tests successful job execution via runner.submit()."""
         mock_compute_resource = MagicMock(spec=ComputeResource)
@@ -91,7 +91,7 @@ class TestScheduleApi(APITestCase):
         mock_compute_resource.save.assert_called_once()
         assert ret_job.status == Job.PENDING
 
-    @patch("scheduler.schedule.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_execute_job_failure(self, mock_get_runner_client):
         """Tests job execution failure handling."""
         mock_runner = MagicMock()
@@ -108,7 +108,7 @@ class TestScheduleApi(APITestCase):
         assert ret_job.status == Job.FAILED
         assert "Compute resource creation or job submission failed" in ret_job.logs
 
-    @patch("scheduler.tasks.update_jobs_statuses.get_runner")
+    @patch("core.services.runners.runner.Runner.get")
     def test_job_runtime_limit(self, get_runner):
         """Tests job runtime limit enforcement.
